# internal/docs/openapi.yaml
openapi: 3.0.3
info:
  title: chaosbox
  description: >
    Small HTTP service used for infra demos and testing: health/version,
    file cat, a counter (in-memory or Redis), CPU/memory/disk load
    simulators, peer membership tracking, and Prometheus metrics. Load
    actions fan out (single hop, best-effort) to known peers.
  version: "0.1.0"
servers:
  - url: /
...
paths:
...
  /load/cpu/start:
    post:
      tags: [load]
      summary: Start a pulsing CPU load (also triggers this on known peers)
      parameters:
        - $ref: "#/components/parameters/MinPct"
        - $ref: "#/components/parameters/MaxPct"
        - $ref: "#/components/parameters/Period"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: object
                properties:
                  status: { type: string, example: started }
                  kind: { type: string, example: cpu }
                  min_pct: { type: number }
                  max_pct: { type: number }
                  period: { type: integer, description: seconds }
...
  /load/mem/start:
    post:
      tags: [load]
      summary: Start a pulsing memory load (also triggers this on known peers)
      parameters:
        - $ref: "#/components/parameters/MemMB"
        - $ref: "#/components/parameters/Period"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: object
                properties:
                  status: { type: string, example: started }
                  kind: { type: string, example: mem }
                  mb: { type: integer, description: peak MB held }
                  period: { type: integer, description: seconds }
        "500":
          description: could not determine available memory
...
  /load/disk/start:
    post:
      tags: [load]
      summary: Start a pulsing disk IO load (also triggers this on known peers)
      parameters:
        - $ref: "#/components/parameters/DiskMB"
        - $ref: "#/components/parameters/Period"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: object
                properties:
                  status: { type: string, example: started }
                  kind: { type: string, example: disk }
                  data: { type: string, description: data directory }
                  file: { type: string, description: load file path }
                  mb: { type: integer }
                  period: { type: integer, description: seconds }
        "500":
          description: disk IO setup failed
...
  /load/all/start:
    post:
      tags: [load]
      summary: Start CPU, memory, and disk load together (also triggers this on known peers)
      parameters:
        - $ref: "#/components/parameters/MinPct"
        - $ref: "#/components/parameters/MaxPct"
        - name: mb
          in: query
          description: memory + disk MB (defaults to 60% available memory / 64MB disk)
          schema: { type: integer }
        - $ref: "#/components/parameters/Period"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: object
                properties:
                  status: { type: string, example: started }
                  kind: { type: string, example: all }
                  min_pct: { type: number }
                  max_pct: { type: number }
                  mem_mb: { type: integer }
                  disk_mb: { type: integer }
                  data: { type: string }
                  file: { type: string }
                  period: { type: integer, description: seconds }
        "500":
          description: load setup failed
...
components:
  parameters:
    MinPct:
      name: min_pct
      in: query
      description: minimum CPU percent (default 60)
      schema: { type: number }
    MaxPct:
      name: max_pct
      in: query
      description: maximum CPU percent (default 80)
      schema: { type: number }
    MemMB:
      name: mb
      in: query
      description: peak memory MB (default 60% of available memory)
      schema: { type: integer }
    DiskMB:
      name: mb
      in: query
      description: disk load file size in MB (default 64)
      schema: { type: integer }
    Period:
      name: period
      in: query
      description: pulse period in seconds (default 10)
      schema: { type: integer }
...